Remove enabling DEPTH_TEST and CULL_FACE
Not too sure this will work out but currently, DEPTH_TEST
might a cause sprites to not get drawn when two or more
share the same coordinates (of course). Since all the sprites
have transparent parts, this is generally unwanted.

And everything should be drawn in the right order already.
So let's just remove DEPTH_TEST and see if we can get away
with it ;)

And CULL_FACE is collateral damage. Won't need that either
probably.